refactor: 타입 단언 지우기, useMateCreate 훅 useeffect 삭제(#117)
