+++ src/pages/match/create/create.tsx
@@ -1,14 +1,14 @@
-import { isInvalidMatchId } from '@pages/match/constants/matching';
 import ButtonSection from '@pages/match/create/components/button-section';
-import Header from '@pages/match/create/components/header';
 import MatchCardSection from '@pages/match/create/components/match-card-section';
+import MatchGuideSection from '@pages/match/create/components/match-guide-section';
 import useMatchCreate from '@pages/match/hooks/use-mate-create';
+import { isInvalidMatchId } from '@pages/match/utils/match-validators';
 import { ROUTES } from '@routes/routes-config';
 import { Navigate, useParams, useSearchParams } from 'react-router-dom';
 
 const Create = () => {
   const [searchParams] = useSearchParams();
-  const matchType = (searchParams.get('type') as 'single' | 'group') || 'single';
+  const matchType = searchParams.get('type') === 'group' ? 'group' : 'single';
   const { matchId } = useParams();
   const numericMatchId = Number(matchId) || 1;
   const { matchData } = useMatchCreate(numericMatchId, matchType);
@@ -20,7 +20,7 @@ const Create = () => {
   return (
     <div className="h-full flex-col-between gap-[2.4rem] px-[1.6rem] pt-[9.6rem]">
       <main className="w-full flex-col-center gap-[4rem]">
-        <Header nickname={matchData?.nickname || ''} />
+        <MatchGuideSection nickname={matchData?.nickname || ''} />
         {matchData && <MatchCardSection matchData={matchData} />}
       </main>
       <ButtonSection matchType={matchType} />
