+++ src/pages/match/utils/match-validators.ts
@@ -1,0 +1,3 @@
+export const isInvalidMatchId = (id: string | undefined): boolean => {
+  return !id || Number.isNaN(Number(id));
+};
